Dodao sam mnogo toga  - Ubacio sam stringove.  - Klasa se lepo iscrtava, a nove fieldove i metode dodajete duplim klikom na njihovo podruchje na klasi. Editujete ih duplim klikom na njih, a brishete tako shto zapishete prazan string i njima.  - I relacije imaju njihove stringove. (mada oni nemaju smisla za generalizaciju i comment link, ali chisto su tu za testiranje) Isto ih tako editujete, a srednji string se mozhe dragovati  - Sada se relacije dodaju drag-n-dropom u smeru od_chega - do_chega. Mnogo je bolje nego klik na jedno pa na drugo.
Mislim da je to to.. U principu je sve spremno za spamovanje interfejsa, agregacija i drugih dhidha-bidha. Josh treba staviti da je comment JTextArea, otom potom.

+++ trunk/RAF UML/edu/raf/uml/model/UMLCommentBox.java
@@ -31,12 +31,12 @@ public class UMLCommentBox extends UMLBox {
     }
 
     @Override
-    public int calculateMinHeight() {
+    public double calculateMinHeight() {
         return 100;
     }
 
     @Override
-    public int calculateMinWidth() {
+    public double calculateMinWidth() {
         return 100;
     }
